don't version jbof.set_mgmt_ip
