+++ src/middlewared/middlewared/api/v25_10_0/jbof.py
@@ -1,17 +1,13 @@
-import ipaddress
 from typing import Literal
 
-from pydantic import Secret, Field, field_validator
+from pydantic import Secret
 
-from middlewared.api.base import (
-    BaseModel, IPvAnyAddress, Excluded, excluded_field, NotRequired, ForUpdateMetaclass, IPv4Address, IPv6Address
-)
+from middlewared.api.base import BaseModel, IPvAnyAddress, Excluded, excluded_field, NotRequired, ForUpdateMetaclass
 
 
 __all__ = [
     'JBOFEntry', 'JBOFCreateArgs', 'JBOFCreateResult', 'JBOFDeleteArgs', 'JBOFDeleteResult', 'JBOFLicensedArgs',
-    'JBOFLicensedResult', 'JBOFReapplyConfigArgs', 'JBOFReapplyConfigResult', 'JBOFSetMgmtIPArgs',
-    'JBOFSetMgmtIPResult', 'JBOFUpdateArgs', 'JBOFUpdateResult',
+    'JBOFLicensedResult', 'JBOFReapplyConfigArgs', 'JBOFReapplyConfigResult', 'JBOFUpdateArgs', 'JBOFUpdateResult',
 ]
 
 
@@ -31,37 +27,6 @@ class JBOFEntry(BaseModel):
 
 class JBOFCreate(JBOFEntry):
     id: Excluded = excluded_field()
-
-
-class StaticIPv4Address(BaseModel):
-    address: IPv4Address = NotRequired
-    netmask: str = NotRequired
-    gateway: IPv4Address = NotRequired
-
-    @field_validator('netmask')
-    @classmethod
-    def validate_netmask(cls, value: str) -> str:
-        if value.isdigit():
-            raise ValueError('Please specify expanded netmask, e.g. 255.255.255.128')
-
-        try:
-            ipaddress.ip_network(f'1.1.1.1/{value}', strict=False)
-        except ValueError:
-            raise ValueError('Not a valid netmask')
-
-
-class StaticIPv6Address(BaseModel):
-    address: IPv6Address = NotRequired
-    prefixlen: int = Field(ge=1, le=64, default=NotRequired)
-
-
-class JBOFSetMgmtIPIOMNetwork(BaseModel):
-    dhcp: bool = NotRequired
-    fqdn: str = NotRequired
-    hostname: str = NotRequired
-    ipv4_static_addresses: list[StaticIPv4Address] | None = None
-    ipv6_static_addresses: list[StaticIPv6Address] | None = None
-    nameservers: list[IPvAnyAddress] | None = None
 
 
 class JBOFUpdate(JBOFCreate, metaclass=ForUpdateMetaclass):
@@ -99,19 +64,6 @@ class JBOFReapplyConfigArgs(BaseModel):
 
 
 class JBOFReapplyConfigResult(BaseModel):
-    result: None
-
-
-class JBOFSetMgmtIPArgs(BaseModel):
-    id: int
-    iom: Literal['IOM1', 'IOM2']
-    iom_network: JBOFSetMgmtIPIOMNetwork = Field(default_factory=JBOFSetMgmtIPIOMNetwork)
-    ethindex: int = 1
-    force: bool = False
-    check: bool = True
-
-
-class JBOFSetMgmtIPResult(BaseModel):
     result: None
